fix double-sync
if the preamble already exists, injecting the preamble would evaluate
the gemfile, and recursively load the plugin (and plugins.rb).
avoid the problems of that by only adding at_exit hooks in multilock.rb,
which will only be required once. the plugin hook needs to be more
carefully considered, because they'll be reset multiple times -
so avoid the re-entrant load in the first place